update loop_var and add conditional to config includes
This change introduces two updates the config includes. First it changes
the loop_var from item to line to avoid name collisions.  Second it adds
a conditional to filter out the "end" config line to avoid leaving
config mode which is handled in the task implementation.

This change also adds the ansible_network_os check into the load_config
functionto handle a case where the function is directly implemented as a
task in playbooks.